requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
numpy>=1.24.0
pytz>=2023.3
google-generativeai>=0.3.0
protobuf~=5.28.3
scipy>=1.11.0